Complete implementation of seek, tell and metadata callbacks
This allows the encoder to seek back to the beginning of the file in
order to write the metadata headers containing things like the hash
and number of samples. Only seek and tell are required for this.

The metadata callback just allows the user to receive a copy of the
metadata. Currently this will arrive as cdata which is not very useful
but it is implemented anyway for completeness.